backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
Flask-SQLAlchemy==3.0.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
markdown==3.5.2
psycopg2-binary==2.9.10
supabase==2.11.0